Refactor env switching for Angular client
