+++ apps/cash-compass/src/app/shared/service/category.service.ts
@@ -2,7 +2,8 @@ import { Injectable } from '@angular/core';
 import {Observable } from "rxjs";
 import {HttpClient} from "@angular/common/http";
 import {Category} from "../model/category.model";
-import { environment } from '../../../environemnts/environment.dev';
+import { environment } from "../../../environments/environment";
+
 const { apiUrl } = environment;
 
 @Injectable({
