+++ apps/cash-compass/src/app/shared/service/expense.service.ts
@@ -2,7 +2,7 @@ import {Injectable} from '@angular/core';
 import {HttpClient} from "@angular/common/http";
 import {BehaviorSubject, map, Observable, tap} from "rxjs";
 import {Expense} from "../model/expense.model";
-import { environment } from '../../../environemnts/environment.dev';
+import { environment } from '../../../environments/environment';
 
 const { apiUrl} = environment;
