+++ apps/cash-compass/src/environments/environment.ts
@@ -1,0 +1,4 @@
+export const environment = {
+  mode: 'development',
+  apiUrl: 'http://localhost:3000/api',
+};
